refactor: replace epicShortDto and epicFullDto with epicResultDto. Do checkstyle
