+++ src/main/java/ru/mssecondteam/taskservice/controller/EpicController.java
@@ -8,15 +8,14 @@ import org.springframework.http.HttpStatus;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import ru.mssecondteam.taskservice.dto.TaskFullDto;
-import ru.mssecondteam.taskservice.dto.epic.dto.EpicFullDto;
-import ru.mssecondteam.taskservice.dto.epic.dto.EpicShortDto;
+import ru.mssecondteam.taskservice.dto.epic.dto.EpicResponseDto;
 import ru.mssecondteam.taskservice.dto.epic.dto.EpicUpdateRequest;
 import ru.mssecondteam.taskservice.dto.epic.dto.NewEpicRequest;
 import ru.mssecondteam.taskservice.mapper.EpicMapper;
 import ru.mssecondteam.taskservice.mapper.TaskMapper;
 import ru.mssecondteam.taskservice.model.Epic;
 import ru.mssecondteam.taskservice.model.Task;
-import ru.mssecondteam.taskservice.service.TaskService;
+import ru.mssecondteam.taskservice.service.EpicService;
 
 @RestController
 @RequestMapping("/epics")
@@ -24,7 +23,7 @@ import ru.mssecondteam.taskservice.service.TaskService;
 @Validated
 @Slf4j
 public class EpicController {
-    private final TaskService taskService;
+    private final EpicService epicService;
 
     private final TaskMapper taskMapper;
 
@@ -32,19 +31,19 @@ public class EpicController {
 
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
-    public EpicShortDto createEpic(@RequestBody @Valid NewEpicRequest newEpic) {
+    public EpicResponseDto createEpic(@RequestBody @Valid NewEpicRequest newEpic) {
         log.debug("Creating epic '{}'", newEpic.title());
         final Epic epic = epicMapper.toEpicModel(newEpic);
-        final Epic createdEpic = taskService.createEpic(epic);
-        return epicMapper.toEpicShortDto(createdEpic);
+        final Epic createdEpic = epicService.createEpic(epic);
+        return epicMapper.toEpicResponseDto(createdEpic);
     }
 
     @PatchMapping("/{epicId}")
-    public EpicShortDto updateEpic(@PathVariable @Positive Long epicId,
-                                   @RequestBody @Valid EpicUpdateRequest updateRequest) {
+    public EpicResponseDto updateEpic(@PathVariable @Positive Long epicId,
+                                      @RequestBody @Valid EpicUpdateRequest updateRequest) {
         log.debug("Updating epic with id '{}'", epicId);
-        final Epic updatedEpic = taskService.updateEpic(epicId, updateRequest);
-        return epicMapper.toEpicShortDto(updatedEpic);
+        final Epic updatedEpic = epicService.updateEpic(epicId, updateRequest);
+        return epicMapper.toEpicResponseDto(updatedEpic);
     }
 
     @PatchMapping("/{epicId}/tasks/{taskId}")
@@ -52,23 +51,23 @@ public class EpicController {
                                      @PathVariable @Positive Long epicId,
                                      @PathVariable @Positive Long taskId) {
         log.debug("Adding task with id '{}' to epic with id '{}' by user with id '{}'", taskId, epicId, userId);
-        final Task addedToEpicTask = taskService.addTaskToEpic(userId, epicId, taskId);
+        final Task addedToEpicTask = epicService.addTaskToEpic(userId, epicId, taskId);
         return taskMapper.toTaskFullDto(addedToEpicTask);
     }
 
-    @PatchMapping("/{epicId}/tasks/{taskId}/delete")
+    @DeleteMapping("/{epicId}/tasks/{taskId}/delete")
     public TaskFullDto deleteTaskFromEpic(@RequestHeader("X-User-Id") @Positive Long userId,
                                           @PathVariable @Positive Long epicId,
                                           @PathVariable @Positive Long taskId) {
         log.debug("Deleting task with id '{}' from epic with id '{}' by user with id '{}'", taskId, epicId, userId);
-        final Task deletedFromEpicTask = taskService.deleteTaskFromEpic(userId, epicId, taskId);
+        final Task deletedFromEpicTask = epicService.deleteTaskFromEpic(userId, epicId, taskId);
         return taskMapper.toTaskFullDto(deletedFromEpicTask);
     }
 
     @GetMapping("/{epicId}")
-    public EpicFullDto findEpicById(@PathVariable @Positive Long epicId) {
+    public EpicResponseDto findEpicById(@PathVariable @Positive Long epicId) {
         log.debug("Retrieving Epic with id '{}'", epicId);
-        final Epic epic = taskService.findEpicById(epicId);
-        return epicMapper.toEpicFullDto(epic);
+        final Epic epic = epicService.findEpicById(epicId);
+        return epicMapper.toEpicResponseDto(epic);
     }
 }
