+++ src/main/java/ru/mssecondteam/taskservice/repository/epic/EpicRepository.java
@@ -1,4 +1,4 @@
-package ru.mssecondteam.taskservice.repository;
+package ru.mssecondteam.taskservice.repository.epic;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.mssecondteam.taskservice.model.Epic;
