+++ src/main/java/ru/mssecondteam/taskservice/service/TaskService.java
@@ -19,14 +19,4 @@ public interface TaskService {
     List<Task> searchTasks(Integer page, Integer size, TaskSearchFilter searchFilter);
 
     void deleteTaskById(Long taskId, Long userId);
-
-    Epic createEpic(Epic epic);
-
-    Epic updateEpic(Long epicId, EpicUpdateRequest updateRequest);
-
-    Task addTaskToEpic(Long userId, Long epicId, Long taskId);
-
-    Task deleteTaskFromEpic(Long userId, Long epicId, Long taskId);
-
-    Epic findEpicById(Long epicId);
 }
